♻️ refactor(ui): introduce shared ui layer with StyledBtn and CssTransition
Move StyledBtn and CssTransition out of their domain-specific locations
into a shared ui/ layer to eliminate layer violations and enable reuse
across generic components (Drawer, Dropdown) and domain components
(AudioPlayer) without creating import coupling.

- Add src/ui/StyledBtn.ts as a unified button atom for all player buttons
- Move src/components/CssTransition.tsx to src/ui/CssTransition.tsx
- Remove Button/StyledBtn.ts (superseded by ui/StyledBtn)
- Update all consumers to import from @/ui/*
- Add @/ui alias to tsconfig, vite, and vitest configs
