+++ package/vitest.config.ts
@@ -10,6 +10,7 @@ export default defineConfig({
       "@/hooks": path.resolve(__dirname, "./src/hooks"),
       "@/utils": path.resolve(__dirname, "./src/utils"),
       "@/styles": path.resolve(__dirname, "./src/styles"),
+      "@/ui": path.resolve(__dirname, "./src/ui"),
     },
   },
   test: {
